Some changes for QNAP code signing and anti-tampering
1. Simplify csv file format
2. Do verification after building with code signing support automatically
